split code generation into separated files
