+++ dan_layer/engine/tests/macros/src/lib.rs
@@ -1,11 +1,11 @@
 mod ast;
-mod template;
+mod output;
 
 use proc_macro::TokenStream;
 
 #[proc_macro]
 pub fn template(input: TokenStream) -> TokenStream {
-    template::generate_template_output(proc_macro2::TokenStream::from(input))
+    output::template::generate_template(proc_macro2::TokenStream::from(input))
         .unwrap_or_else(|err| err.to_compile_error())
         .into()
 }
